+++ wxc_sdk/telephony/__init__.py
@@ -206,6 +206,9 @@ class DestinationType(str, Enum):
 class LocationAndNumbers(ApiModel):
     @field_validator('phone_number', mode='before')
     def e164(cls, v):
+        """
+        :meta private:
+        """
         return plus1(v)
 
     location_name: str